Fix local filesystem tests so they all work, and all clean up properly their temp files, and also test the scenario of root folder that doesn't yet exist before opening the Local FS.

+++ filesystem/src/test/java/org/frankframework/filesystem/LocalFileSystemTestHelper.java
@@ -21,7 +21,9 @@ public class LocalFileSystemTestHelper implements IFileSystemTestHelper {
 
 	@Override
 	public void setUp() throws Exception {
-		FileUtils.cleanDirectory(folder.toFile());
+		if (Files.exists(folder)) {
+			FileUtils.cleanDirectory(folder.toFile());
+		}
 	}
 
 	@Override
